Double in kamheat

+++ driver_kamheat.h
@@ -13,8 +13,8 @@
 struct Kamheat: Driver
 {
   Kamheat(std::string key = "") : Driver(std::string("kamheat"), key) {};
-  virtual esphome::optional<std::map<std::string, float>> get_values(std::vector<unsigned char> &telegram) override {
-    std::map<std::string, float> ret_val{};
+  virtual esphome::optional<std::map<std::string, double>> get_values(std::vector<unsigned char> &telegram) override {
+    std::map<std::string, double> ret_val{};
 
     add_to_map(ret_val, "total_energy_consumption_gj", this->get_current_heating_GJ(telegram));
     add_to_map(ret_val, "total_forward_energy_m3c", this->get_forward_energy_m3c(telegram));
@@ -33,8 +33,8 @@ struct Kamheat: Driver
   };
 
 private:
-  esphome::optional<float> get_current_heating_GJ(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_current_heating_GJ(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x31) && (telegram[19] == 0x79)) {  //short frame (L=0x31) and 0x79 tpl-ci-field (EN 13757-3 Application Layer with Compact frame (no tplh)) 
       i+=2;
@@ -46,8 +46,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_forward_energy_m3c(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_forward_energy_m3c(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x40) && (telegram[19] == 0x78)) {  //longer frame (L=0x40) and 0x78 tpl-ci-field      
       ret_val = (((uint32_t)telegram[i+13] << 24) + ((uint32_t)telegram[i+12] << 16) + ((uint32_t)telegram[i+11] << 8) + (uint32_t)telegram[i+10]);
@@ -55,8 +55,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_return_energy_m3c(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_return_energy_m3c(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x40) && (telegram[19] == 0x78)) {  //longer frame (L=0x40) and 0x78 tpl-ci-field   
       ret_val = (((uint32_t)telegram[i+20] << 24) + ((uint32_t)telegram[i+19] << 16) + ((uint32_t)telegram[i+18] << 8) + (uint32_t)telegram[i+17]);
@@ -64,8 +64,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_total_volume_m3(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_total_volume_m3(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x31) && (telegram[19] == 0x79)){  //short frame (L=0x31) and 0x79 tpl-ci-field (EN 13757-3 Application Layer with Compact frame (no tplh))
       i-=6;
@@ -77,8 +77,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_volume_flow_lh(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_volume_flow_lh(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x31) && (telegram[19] == 0x79)) {  //short frame (L=0x31) and 0x79 tpl-ci-field (EN 13757-3 Application Layer with Compact frame (no tplh))
       i-=11;
@@ -90,8 +90,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_temperature_inlet_C(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_temperature_inlet_C(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x31) && (telegram[19] == 0x79)) {  //short frame (L=0x31) and 0x79 tpl-ci-field (EN 13757-3 Application Layer with Compact frame (no tplh))
       i-=13;
@@ -103,8 +103,8 @@ private:
     return ret_val;
   };
 
-  esphome::optional<float> get_temperature_outlet_C(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_temperature_outlet_C(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 19;
     if ((telegram[0] == 0x31) && (telegram[19] == 0x79)) {  //short frame (L=0x31) and 0x79 tpl-ci-field (EN 13757-3 Application Layer with Compact frame (no tplh))
       i-=15;
